Estetyczne zmiany wprowadzane do projektow. BusinessLogic, DatabaseAPI, HelperProject oraz MobileApp

+++ BusinessLogic/CartBusinessLogic.cs
@@ -2,12 +2,7 @@
 using DatabaseAPI.Models.CinemaMovie;
 using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.Sockets;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace BusinessLogic
 {
@@ -26,7 +21,7 @@ namespace BusinessLogic
         {
             if (httpContext.Session.GetString("SessionId") == null)
             {
-                if (!string.IsNullOrWhiteSpace(httpContext.User.Identity.Name))
+                if (!string.IsNullOrWhiteSpace(httpContext.User.Identity?.Name))
                 {
                     httpContext.Session.SetString("SessionId", httpContext.User.Identity.Name);
                 }
@@ -36,7 +31,7 @@ namespace BusinessLogic
                     httpContext.Session.SetString("SessionId", tempSessionId.ToString());
                 }
             }
-            return httpContext.Session.GetString("SessionId").ToString();
+            return httpContext.Session.GetString("SessionId")?.ToString();
         }
 
         public string GetSessionId()
@@ -46,31 +41,33 @@ namespace BusinessLogic
 
         public async Task RemoveFromCart(CartElement element)
         {
-            dbContext.CartElement.Remove(element);
+            dbContext.CartElement?.Remove(element);
             await dbContext.SaveChangesAsync();
         }
 
         public async Task ClearCart(CartElement[] elements)
         {
-            dbContext.CartElement.RemoveRange(elements);
+            dbContext.CartElement?.RemoveRange(elements);
             await dbContext.SaveChangesAsync();
         }
 
         public async Task ClearCoupon(int? couponId)
         {
             if (couponId == null) return;
-            var couponCartElement = await dbContext.CartElement.FirstOrDefaultAsync(ele => ele.SessionId == this.SessionId &&  ele.CouponId == couponId);
+            var couponCartElement = await dbContext.CartElement?.FirstAsync(ele => ele.SessionId == this.SessionId && ele.CouponId == couponId);
+            if (couponCartElement == null)
+            {
+                return;
+            }
             dbContext.CartElement.Remove(couponCartElement);
             await dbContext.SaveChangesAsync();
         }
 
-        // I need to add logic to check if ticket was taken.
-        public void AddToCart(ScreeningSeat screeningSeat)
+        public async Task AddToCart(ScreeningSeat screeningSeat)
         {
-            dbContext.CartElement.Add(NewCartElement(screeningSeat));
-            dbContext.SaveChanges();
+            await dbContext.CartElement.AddAsync(NewCartElement(screeningSeat));
+            await dbContext.SaveChangesAsync();
         }
-        // I need to check if I want to edit ticket added to a cart or I prefer to delete it and then add new one.
 
         public async Task<List<CartElement>> GetCartElements()
         {
@@ -88,10 +85,10 @@ namespace BusinessLogic
                 .ThenInclude(item => item.Screening).ThenInclude(item => item.Movie).SumAsync(item => item.ScreeningSeat.Screening.Movie.TicketPrice) ?? decimal.Zero;
         }
 
-        public Coupon CheckCoupon()
+        public async Task<Coupon> CheckCoupon()
         {
-            var cartEleCoupon = dbContext.CartElement.Where(item => item.SessionId == this.SessionId && item.CouponId != null).Include(item => item.Coupon).ToList();
-            if (cartEleCoupon.Any() )
+            var cartEleCoupon = await dbContext.CartElement.Where(item => item.SessionId == this.SessionId && item.CouponId != null).Include(item => item.Coupon).ToListAsync();
+            if (cartEleCoupon.Any())
             {
                 return cartEleCoupon.First().Coupon;
             }
